requirements.txt
absl-py==2.1.0
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy==1.3.0
cycler==0.12.1
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.56.0
fsspec==2025.2.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
grpcio==1.70.0
huggingface-hub==0.29.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2 @ file:///croot/jinja2_1737760107953/work
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.9.4
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.24.4
packaging==24.2
pillow @ file:///croot/pillow_1738010226202/work
protobuf==5.29.3
pyparsing==3.2.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1730999120400/work
safetensors==0.5.3
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.15
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
urllib3 @ file:///croot/urllib3_1737133630106/work
Werkzeug==3.1.3
zipp==3.21.0
